Attempt reconnection when sending state to offline device
This change modifies the `sendState` function in `WebsocketClient` to automatically trigger a reconnection attempt if the device is not connected.

Additionally, the brightness slider in the device list is now always enabled, removing the dependency on the device's online status to allow state changes to be sent even when offline, which will now trigger the reconnection.